Final commit
Not very happy with my performance for the first course. I had expected to fly right through the course with my past experience playing with different languages. However I did not expect such a heavy focus on discrete math to complete in under the specified time limit for the function. I intend to study discrete math before I continue the next courses. Let this be an example for my progress as I move through the courses.